Found memory leak
I don't know why, but calling elapsedTimeSlider.setValue(elapsedTime)
is the source of my problem. As far as I can tell, there is nothing I
can do about it. I've replicated the issue with a minimal amount of
code, and the problem persists. Simply calling the built in setValue
method eats memory. I don't know if this is an openjfx-11 problem, an
operating system problem, a window manager problem, or if there is
some way to fix this I've not yet considered. As it stands now, an
elapsed time slider is being pushed to the back of the priority list,
unfortunately.